Breakout charm API into common.
