+++ api/charms/client_test.go
@@ -5,7 +5,6 @@ package charms_test
 
 import (
 	"github.com/golang/mock/gomock"
-	charm "github.com/juju/charm/v8"
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
@@ -39,65 +38,4 @@ func (s *charmsMockSuite) TestIsMeteredFalse(c *gc.C) {
 	got, err := client.IsMetered(url)
 	c.Assert(err, gc.IsNil)
 	c.Assert(got, jc.IsTrue)
-}
-
-func (s *charmsMockSuite) TestCharmInfo(c *gc.C) {
-	ctrl := gomock.NewController(c)
-	defer ctrl.Finish()
-
-	mockFacadeCaller := basemocks.NewMockFacadeCaller(ctrl)
-
-	url := "local:quantal/dummy-1"
-	args := params.CharmURL{URL: url}
-	info := new(params.Charm)
-
-	params := params.Charm{
-		Revision: 1,
-		URL:      url,
-		Config: map[string]params.CharmOption{
-			"config": {
-				Type:        "type",
-				Description: "config-type option",
-			},
-		},
-		LXDProfile: &params.CharmLXDProfile{
-			Description: "LXDProfile",
-			Devices: map[string]map[string]string{
-				"tun": {
-					"path": "/dev/net/tun",
-					"type": "unix-char",
-				},
-			},
-		},
-	}
-
-	mockFacadeCaller.EXPECT().FacadeCall("CharmInfo", args, info).SetArg(2, params).Return(nil)
-
-	client := charms.NewClientWithFacade(mockFacadeCaller)
-	got, err := client.CharmInfo(url)
-	c.Assert(err, gc.IsNil)
-
-	want := &charms.CharmInfo{
-		Revision: 1,
-		URL:      url,
-		Config: &charm.Config{
-			Options: map[string]charm.Option{
-				"config": {
-					Type:        "type",
-					Description: "config-type option",
-				},
-			},
-		},
-		LXDProfile: &charm.LXDProfile{
-			Description: "LXDProfile",
-			Config:      map[string]string{},
-			Devices: map[string]map[string]string{
-				"tun": {
-					"path": "/dev/net/tun",
-					"type": "unix-char",
-				},
-			},
-		},
-	}
-	c.Assert(got, gc.DeepEquals, want)
 }
