+++ apiserver/facades/client/charms/client.go
@@ -5,11 +5,11 @@ package charms
 
 import (
 	"github.com/juju/charm/v8"
-	"github.com/juju/charm/v8/resource"
 	"github.com/juju/collections/set"
 	"github.com/juju/errors"
 	"github.com/juju/names/v4"
 
+	charmscommon "github.com/juju/juju/apiserver/common/charms"
 	apiservererrors "github.com/juju/juju/apiserver/errors"
 	"github.com/juju/juju/apiserver/facade"
 	"github.com/juju/juju/apiserver/params"
@@ -26,6 +26,7 @@ type backend interface {
 // API implements the charms interface and is the concrete
 // implementation of the API end point.
 type API struct {
+	*charmscommon.CharmsAPI
 	authorizer facade.Authorizer
 	backend    backend
 }
@@ -54,7 +55,13 @@ func NewFacade(ctx facade.Context) (*API, error) {
 		return nil, errors.Trace(err)
 	}
 
+	commonCharmsAPI, err := charmscommon.NewCharmsAPI(st, authorizer)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+
 	return &API{
+		CharmsAPI:  commonCharmsAPI,
 		authorizer: authorizer,
 		backend:    getState(st, m),
 	}, nil
@@ -69,39 +76,6 @@ type stateShim struct {
 
 var getState = func(st *state.State, m *state.Model) backend {
 	return stateShim{st, m}
-}
-
-// CharmInfo returns information about the requested charm.
-// NOTE: thumper 2016-06-29, this is not a bulk call and probably should be.
-func (a *API) CharmInfo(args params.CharmURL) (params.Charm, error) {
-	if err := a.checkCanRead(); err != nil {
-		return params.Charm{}, errors.Trace(err)
-	}
-
-	curl, err := charm.ParseURL(args.URL)
-	if err != nil {
-		return params.Charm{}, errors.Trace(err)
-	}
-	aCharm, err := a.backend.Charm(curl)
-	if err != nil {
-		return params.Charm{}, errors.Trace(err)
-	}
-	info := params.Charm{
-		Revision: aCharm.Revision(),
-		URL:      curl.String(),
-		Config:   params.ToCharmOptionMap(aCharm.Config()),
-		Meta:     convertCharmMeta(aCharm.Meta()),
-		Actions:  convertCharmActions(aCharm.Actions()),
-		Metrics:  convertCharmMetrics(aCharm.Metrics()),
-	}
-
-	// we don't need to check that this is a charm.LXDProfiler, as we can
-	// state that the function exists.
-	if profile := aCharm.LXDProfile(); profile != nil && !profile.Empty() {
-		info.LXDProfile = convertCharmLXDProfile(profile)
-	}
-
-	return info, nil
 }
 
 // List returns a list of charm URLs currently in the state.
@@ -150,217 +124,4 @@ func (a *API) IsMetered(args params.CharmURL) (params.IsMeteredResult, error) {
 		return params.IsMeteredResult{Metered: true}, nil
 	}
 	return params.IsMeteredResult{Metered: false}, nil
-}
-
-func convertCharmMeta(meta *charm.Meta) *params.CharmMeta {
-	if meta == nil {
-		return nil
-	}
-	return &params.CharmMeta{
-		Name:           meta.Name,
-		Summary:        meta.Summary,
-		Description:    meta.Description,
-		Subordinate:    meta.Subordinate,
-		Provides:       convertCharmRelationMap(meta.Provides),
-		Requires:       convertCharmRelationMap(meta.Requires),
-		Peers:          convertCharmRelationMap(meta.Peers),
-		ExtraBindings:  convertCharmExtraBindingMap(meta.ExtraBindings),
-		Categories:     meta.Categories,
-		Tags:           meta.Tags,
-		Series:         meta.Series,
-		Storage:        convertCharmStorageMap(meta.Storage),
-		PayloadClasses: convertCharmPayloadClassMap(meta.PayloadClasses),
-		Resources:      convertCharmResourceMetaMap(meta.Resources),
-		Terms:          meta.Terms,
-		MinJujuVersion: meta.MinJujuVersion.String(),
-	}
-}
-
-func convertCharmRelationMap(relations map[string]charm.Relation) map[string]params.CharmRelation {
-	if len(relations) == 0 {
-		return nil
-	}
-	result := make(map[string]params.CharmRelation)
-	for key, value := range relations {
-		result[key] = convertCharmRelation(value)
-	}
-	return result
-}
-
-func convertCharmRelation(relation charm.Relation) params.CharmRelation {
-	return params.CharmRelation{
-		Name:      relation.Name,
-		Role:      string(relation.Role),
-		Interface: relation.Interface,
-		Optional:  relation.Optional,
-		Limit:     relation.Limit,
-		Scope:     string(relation.Scope),
-	}
-}
-
-func convertCharmStorageMap(storage map[string]charm.Storage) map[string]params.CharmStorage {
-	if len(storage) == 0 {
-		return nil
-	}
-	result := make(map[string]params.CharmStorage)
-	for key, value := range storage {
-		result[key] = convertCharmStorage(value)
-	}
-	return result
-}
-
-func convertCharmStorage(storage charm.Storage) params.CharmStorage {
-	return params.CharmStorage{
-		Name:        storage.Name,
-		Description: storage.Description,
-		Type:        string(storage.Type),
-		Shared:      storage.Shared,
-		ReadOnly:    storage.ReadOnly,
-		CountMin:    storage.CountMin,
-		CountMax:    storage.CountMax,
-		MinimumSize: storage.MinimumSize,
-		Location:    storage.Location,
-		Properties:  storage.Properties,
-	}
-}
-
-func convertCharmPayloadClassMap(payload map[string]charm.PayloadClass) map[string]params.CharmPayloadClass {
-	if len(payload) == 0 {
-		return nil
-	}
-	result := make(map[string]params.CharmPayloadClass)
-	for key, value := range payload {
-		result[key] = convertCharmPayloadClass(value)
-	}
-	return result
-}
-
-func convertCharmPayloadClass(payload charm.PayloadClass) params.CharmPayloadClass {
-	return params.CharmPayloadClass{
-		Name: payload.Name,
-		Type: payload.Type,
-	}
-}
-
-func convertCharmResourceMetaMap(resources map[string]resource.Meta) map[string]params.CharmResourceMeta {
-	if len(resources) == 0 {
-		return nil
-	}
-	result := make(map[string]params.CharmResourceMeta)
-	for key, value := range resources {
-		result[key] = convertCharmResourceMeta(value)
-	}
-	return result
-}
-
-func convertCharmResourceMeta(meta resource.Meta) params.CharmResourceMeta {
-	return params.CharmResourceMeta{
-		Name:        meta.Name,
-		Type:        meta.Type.String(),
-		Path:        meta.Path,
-		Description: meta.Description,
-	}
-}
-
-func convertCharmActions(actions *charm.Actions) *params.CharmActions {
-	if actions == nil {
-		return nil
-	}
-	result := &params.CharmActions{
-		ActionSpecs: convertCharmActionSpecMap(actions.ActionSpecs),
-	}
-
-	return result
-}
-
-func convertCharmActionSpecMap(specs map[string]charm.ActionSpec) map[string]params.CharmActionSpec {
-	if len(specs) == 0 {
-		return nil
-	}
-	result := make(map[string]params.CharmActionSpec)
-	for key, value := range specs {
-		result[key] = convertCharmActionSpec(value)
-	}
-	return result
-}
-
-func convertCharmActionSpec(spec charm.ActionSpec) params.CharmActionSpec {
-	return params.CharmActionSpec{
-		Description: spec.Description,
-		Params:      spec.Params,
-	}
-}
-
-func convertCharmMetrics(metrics *charm.Metrics) *params.CharmMetrics {
-	if metrics == nil {
-		return nil
-	}
-	return &params.CharmMetrics{
-		Metrics: convertCharmMetricMap(metrics.Metrics),
-		Plan:    convertCharmPlan(metrics.Plan),
-	}
-}
-
-func convertCharmPlan(plan *charm.Plan) params.CharmPlan {
-	if plan == nil {
-		return params.CharmPlan{Required: false}
-	}
-	return params.CharmPlan{Required: plan.Required}
-}
-
-func convertCharmMetricMap(metrics map[string]charm.Metric) map[string]params.CharmMetric {
-	if len(metrics) == 0 {
-		return nil
-	}
-	result := make(map[string]params.CharmMetric)
-	for key, value := range metrics {
-		result[key] = convertCharmMetric(value)
-	}
-	return result
-}
-
-func convertCharmMetric(metric charm.Metric) params.CharmMetric {
-	return params.CharmMetric{
-		Type:        string(metric.Type),
-		Description: metric.Description,
-	}
-}
-
-func convertCharmExtraBindingMap(bindings map[string]charm.ExtraBinding) map[string]string {
-	if len(bindings) == 0 {
-		return nil
-	}
-	result := make(map[string]string)
-	for key, value := range bindings {
-		result[key] = value.Name
-	}
-	return result
-}
-
-func convertCharmLXDProfile(profile *state.LXDProfile) *params.CharmLXDProfile {
-	return &params.CharmLXDProfile{
-		Description: profile.Description,
-		Config:      convertCharmLXDProfileConfig(profile.Config),
-		Devices:     convertCharmLXDProfileDevices(profile.Devices),
-	}
-}
-
-func convertCharmLXDProfileConfig(config map[string]string) map[string]string {
-	result := map[string]string{}
-	for k, v := range config {
-		result[k] = v
-	}
-	return result
-}
-
-func convertCharmLXDProfileDevices(devices map[string]map[string]string) map[string]map[string]string {
-	result := map[string]map[string]string{}
-	for k, v := range devices {
-		nested := map[string]string{}
-		for nk, nv := range v {
-			nested[nk] = nv
-		}
-		result[k] = nested
-	}
-	return result
 }
