+++ cmd/juju/application/deployer/steps.go
@@ -6,7 +6,7 @@ package deployer
 import (
 	"github.com/juju/romulus"
 
-	apicharms "github.com/juju/juju/api/charms"
+	apicommoncharms "github.com/juju/juju/api/common/charms"
 	"github.com/juju/juju/charmstore"
 )
 
@@ -27,7 +27,7 @@ type DeploymentInfo struct {
 	CharmID         charmstore.CharmID
 	ApplicationName string
 	ModelUUID       string
-	CharmInfo       *apicharms.CharmInfo
+	CharmInfo       *apicommoncharms.CharmInfo
 	ApplicationPlan string
 	Force           bool
 }
